gcp_vision_api_demo/requirement.txt
cachetools              ==  3.1.1
certifi                 ==  2019.6.16
chardet                 ==  3.0.4
google-api-core        ==   1.14.2
google-api-python-client == 1.7.10
google-auth             ==  1.6.3
google-auth-httplib2    ==  0.0.3
googleapis-common-protos == 1.6.0
grpcio                  ==  1.22.0
gunicorn                ==  19.9.0
httplib2                ==  0.13.1
idna                    ==  2.8
oauth2client            ==  4.1.3
pip                     ==  19.2.1
protobuf                ==  3.9.1
pyasn1                  ==  0.4.6
pyasn1-modules          ==  0.2.6
pytz                    ==  2019.2
requests                ==  2.22.0
rsa                     ==  4.0
setuptools              ==  41.0.1
six                     ==  1.12.0
sqlparse                ==  0.3.0
uritemplate             ==  3.0.0
urllib3                 ==  1.25.3
wheel                   ==  0.33.4
google-cloud-storage    ==  1.13.1
google-cloud-core       ==  0.29.0
google-cloud-vision     ==  0.41.0
